+++ examples/Console/Asynchronous/EthanA2/AI/AI_continuous_multi_slot.py
@@ -1,0 +1,2 @@
+
+## EthanA2 did not support this example code