+++ examples/Console/Synchronous/EthanP/DPOT/DPOT_writeByChannel.py
@@ -1,0 +1,73 @@
+'''
+DPOT - DPOT_writeByChannel.py with asynchronous mode.
+
+This example demonstrates how to write digital potentiometer resistance to the specefic channel from EthanP.
+-------------------------------------------------------------------------------------
+Please change correct serial number or IP and port number BEFORE you run example code.
+
+For other examples please check:
+    https://github.com/WPC-Systems-Ltd/WPC_Python_driver_release/tree/main/examples
+See README.md file to get detailed usage of this example.
+
+Copyright (c) 2022-2024 WPC Systems Ltd. All rights reserved.
+'''
+
+## Python
+import time
+
+## WPC
+
+from wpcsys import pywpc
+
+def main():
+    ## Get Python driver version
+    print(f'{pywpc.PKG_FULL_NAME} - Version {pywpc.__version__}')
+
+    ## Create device handle
+    dev = pywpc.EthanP()
+
+    ## Connect to device
+    try:
+        dev.connect("192.168.1.110") ## Depend on your device
+    except Exception as err:
+        pywpc.printGenericError(err)
+        ## Release device handle
+        dev.close()
+        return
+
+    try:
+        ## Parameters setting
+        port = 0 ## Depend on your device
+        channel = 0
+        resistance_ratio = 0.5
+        timeout = 3 ## second
+
+        ## Get firmware model & version
+        driver_info = dev.Sys_getDriverInfo(timeout)
+        print("Model name: " + driver_info[0])
+        print("Firmware version: " + driver_info[-1])
+
+        ## Open DPOT
+        err = dev.DPOT_open(port, timeout)
+        print(f"DPOT_open in port {port}, status: {err}")
+
+        ## Write on the channel
+        err = dev.DPOT_writeByChannel(port, channel, resistance_ratio, timeout)
+        print(f"DPOT_writeByChannel in port {port} channel {channel}, status: {err}")
+
+        ## Close DPOT
+        err = dev.DPOT_close(port, timeout)
+        print(f"DPOT_close in port {port}, status: {err}")
+    except Exception as err:
+        pywpc.printGenericError(err)
+
+    ## Disconnect device
+    dev.disconnect()
+
+    ## Release device handle
+    dev.close()
+
+    return
+
+if __name__ == '__main__':
+    main()